Move AppCaseMetadata jsonobjects to app_schemas
