+++ corehq/apps/app_manager/app_schemas/app_case_metadata.py
@@ -1,6 +1,22 @@
 from __future__ import absolute_import, unicode_literals
 
+import datetime
 import re
+from collections import namedtuple
+
+from django.utils.translation import ugettext_lazy as _
+
+import six
+from jsonobject.base import DefaultProperty
+
+from dimagi.ext.jsonobject import (
+    BooleanProperty,
+    DictProperty,
+    JsonObject,
+    ListProperty,
+    ObjectProperty,
+    StringProperty,
+)
 
 from corehq.apps.app_manager.app_schemas.case_properties import (
     get_parent_type_map,
@@ -14,11 +30,10 @@ from corehq.apps.app_manager.models import (
     FormActionCondition,
     Module,
 )
+from corehq.apps.app_manager.xform import VELLUM_TYPES
 from corehq.apps.data_dictionary.util import get_case_property_description_dict
-from corehq.apps.reports.formdetails.readable import (
-    AppCaseMetadata,
-    FormQuestionResponse,
-)
+from corehq.util.timezones.conversions import PhoneTime
+from corehq.util.timezones.utils import get_timezone_for_request
 
 
 class AppCaseMetadataBuilder(object):
@@ -220,3 +235,319 @@ class _AdvancedFormCaseMetadataBuilder(_BaseFormCaseMetadataBuilder):
             meta.add_opener(self.form.unique_id, action.open_condition)
             if action.close_condition.is_active():
                 meta.add_closer(self.form.unique_id, action.close_condition)
+
+
+class CaseMetaException(Exception):
+    pass
+
+
+class FormQuestionOption(JsonObject):
+    label = StringProperty()
+    value = StringProperty()
+
+
+class FormQuestion(JsonObject):
+    label = StringProperty()
+    translations = DictProperty(exclude_if_none=True)
+    tag = StringProperty()
+    type = StringProperty(choices=list(VELLUM_TYPES))
+    value = StringProperty()
+    repeat = StringProperty()
+    group = StringProperty()
+    options = ListProperty(FormQuestionOption, exclude_if_none=True)
+    calculate = StringProperty()
+    relevant = StringProperty()
+    required = BooleanProperty()
+    comment = StringProperty()
+    setvalue = StringProperty()
+
+    @property
+    def icon(self):
+        try:
+            return VELLUM_TYPES[self.type]['icon']
+        except KeyError:
+            return 'fa fa-question-circle'
+
+    @property
+    def relative_value(self):
+        if self.group:
+            prefix = self.group + '/'
+            if self.value.startswith(prefix):
+                return self.value[len(prefix):]
+        return '/'.join(self.value.split('/')[2:])
+
+    @property
+    def option_values(self):
+        return [o.value for o in self.options]
+
+    @property
+    def editable(self):
+        if not self.type:
+            return True
+        vtype = VELLUM_TYPES[self.type]
+        if 'editable' not in vtype:
+            return False
+        return vtype['editable']
+
+
+class FormQuestionResponse(FormQuestion):
+    response = DefaultProperty()
+    children = ListProperty(lambda: FormQuestionResponse, exclude_if_none=True)
+
+    def get_formatted_response(self):
+        timezone = get_timezone_for_request()
+        if self.type == 'DateTime' and timezone \
+                and isinstance(self.response, datetime.datetime):
+            return (PhoneTime(self.response, timezone).user_time(timezone)
+                    .ui_string())
+        else:
+            return self.response
+
+
+class ConditionalFormQuestionResponse(JsonObject):
+    question = ObjectProperty(FormQuestionResponse)
+    condition = ObjectProperty(FormActionCondition)
+
+
+class QuestionList(JsonObject):
+    questions = ListProperty(FormQuestionResponse)
+
+
+class ConditionList(JsonObject):
+    conditions = ListProperty(FormActionCondition)
+
+
+class CaseFormMeta(JsonObject):
+    form_id = StringProperty()
+    load_questions = ListProperty(ConditionalFormQuestionResponse)
+    save_questions = ListProperty(ConditionalFormQuestionResponse)
+    errors = ListProperty(six.text_type)
+
+
+class CaseDetailMeta(JsonObject):
+    module_id = StringProperty()
+    header = DictProperty()
+    error = StringProperty()
+
+
+class CaseProperty(JsonObject):
+    case_type = StringProperty()
+    name = StringProperty()
+    forms = ListProperty(CaseFormMeta)
+    short_details = ListProperty(CaseDetailMeta)
+    long_details = ListProperty(CaseDetailMeta)
+    has_errors = BooleanProperty()
+    description = StringProperty()
+    is_detail_calculation = BooleanProperty()
+
+    def get_form(self, form_id):
+        try:
+            form = next(form for form in self.forms if form.form_id == form_id)
+        except StopIteration:
+            form = CaseFormMeta(form_id=form_id)
+            self.forms.append(form)
+        return form
+
+    def add_load(self, form_id, question):
+        form = self.get_form(form_id)
+        form.load_questions.append(ConditionalFormQuestionResponse(
+            question=question,
+            condition=None
+        ))
+
+    def add_save(self, form_id, question, condition=None):
+        form = self.get_form(form_id)
+        form.save_questions.append(ConditionalFormQuestionResponse(
+            question=question,
+            condition=(condition if condition and condition.type == 'if' else None)
+        ))
+
+    def add_detail(self, type_, module_id, header, is_detail_calculation=False, error=None):
+        self.is_detail_calculation = is_detail_calculation
+        {
+            "short": self.short_details,
+            "long": self.long_details,
+        }[type_].append(CaseDetailMeta(module_id=module_id, header=header, error=error))
+
+
+class CaseTypeMeta(JsonObject):
+    name = StringProperty(required=True)
+    relationships = DictProperty()  # relationship name -> [child type 1, ...]
+    properties = ListProperty(CaseProperty)  # property -> CaseProperty
+    opened_by = DictProperty(ConditionList)  # form_ids -> [FormActionCondition, ...]
+    closed_by = DictProperty(ConditionList)  # form_ids -> [FormActionCondition, ...]
+    error = StringProperty()
+    has_errors = BooleanProperty()
+
+    # store where this case type gets loaded so that we can look it up more easily later
+    load_properties = DictProperty()  # {"form_id": {"question_path": [CaseProperty, ...]}}
+    save_properties = DictProperty()  # {"form_id": {"question_path": [CaseProperty, ...]}}
+
+    @property
+    def child_types(self):
+        """ A list of all child types
+        """
+        return [child_type for relationship in self.relationships.values() for child_type in relationship]
+
+    def get_property(self, name, allow_parent=False):
+        if not allow_parent:
+            assert '/' not in name, "Add parent properties to the correct case type"
+        try:
+            prop = next(prop for prop in self.properties if prop.name == name)
+        except StopIteration:
+            prop = CaseProperty(name=name, case_type=self.name)
+            self.properties.append(prop)
+        return prop
+
+    def add_opener(self, form_id, condition):
+        openers = self.opened_by.get(form_id, ConditionList())
+        if condition.type == 'if':
+            # only add optional conditions
+            openers.conditions.append(condition)
+        self.opened_by[form_id] = openers
+
+    def add_closer(self, form_id, condition):
+        closers = self.closed_by.get(form_id, ConditionList())
+        if condition.type == 'if':
+            # only add optional conditions
+            closers.conditions.append(condition)
+        self.closed_by[form_id] = closers
+
+    def add_save(self, form_id, question_path, property_):
+        if self.get_save_properties(form_id, question_path):
+            self.save_properties[form_id][question_path].append(property_)
+        else:
+            try:
+                self.save_properties[form_id].update({question_path: [property_]})
+            except KeyError:
+                self.save_properties[form_id] = {question_path: [property_]}
+
+    def add_load(self, form_id, question_path, property_):
+        if self.get_load_properties(form_id, question_path):
+            self.load_properties[form_id][question_path].append(property_)
+        else:
+            try:
+                self.load_properties[form_id].update({question_path: [property_]})
+            except KeyError:
+                self.load_properties[form_id] = {question_path: [property_]}
+
+    def get_load_properties(self, form_id, path):
+        """returns a list of properties which load into a particular form question
+        """
+        return self.load_properties.get(form_id, {}).get(path, [])
+
+    def get_save_properties(self, form_id, path):
+        """returns a list of properties which load into a particular form question
+        """
+        return self.save_properties.get(form_id, {}).get(path, [])
+
+
+LoadSaveProperty = namedtuple('LoadSaveProperty', 'case_type property')
+
+
+class AppCaseMetadata(JsonObject):
+    case_types = ListProperty(CaseTypeMeta)  # case_type -> CaseTypeMeta
+
+    def get_load_properties(self, form_id, path):
+        """gets all case types with a list of properties which load into a form question
+        """
+        return [
+            LoadSaveProperty(case_type.name, prop)
+            for case_type in self.case_types
+            for prop in case_type.get_load_properties(form_id, path)
+        ]
+
+    def get_save_properties(self, form_id, path):
+        """gets all case types with a list of properties which are saved from a form question
+        """
+        return [
+            LoadSaveProperty(case_type.name, prop)
+            for case_type in self.case_types
+            for prop in case_type.get_save_properties(form_id, path)
+        ]
+
+    def get_property_list(self, root_case_type, name):
+        type_ = self.get_type(root_case_type)
+        if '/' in name:
+            # find the case property from the correct case type
+            parent_rel, name = name.split('/', 1)
+            parent_case_types = type_.relationships.get(parent_rel)
+            if parent_case_types:
+                parent_props = [
+                    prop for parent_case_type in parent_case_types
+                    for prop in self.get_property_list(parent_case_type, name)
+                ]
+                if parent_props:
+                    return parent_props
+            else:
+                params = {'case_type': root_case_type, 'relationship': parent_rel}
+                raise CaseMetaException(_(
+                    "Case type '%(case_type)s' has no '%(relationship)s' "
+                    "relationship to any other case type.") % params)
+
+        return [type_.get_property(name)]
+
+    def add_property_load(self, root_case_type, name, form_id, question):
+        try:
+            props = self.get_property_list(root_case_type, name)
+        except CaseMetaException as e:
+            props = [self.add_property_error(root_case_type, name, form_id, str(e))]
+
+        for prop in props:
+            self.get_type(prop.case_type).add_load(form_id, question.value, prop.name)
+            prop.add_load(form_id, question)
+
+    def add_property_save(self, root_case_type, name, form_id, question, condition=None):
+        try:
+            props = self.get_property_list(root_case_type, name)
+        except CaseMetaException as e:
+            props = [self.add_property_error(root_case_type, name, form_id, str(e))]
+
+        for prop in props:
+            self.get_type(prop.case_type).add_save(form_id, question.value, prop.name)
+            prop.add_save(form_id, question, condition)
+
+    def add_property_error(self, case_type, case_property, form_id, message):
+        prop = self.get_error_property(case_type, case_property)
+        prop.has_errors = True
+        if form_id is not None:
+            form = prop.get_form(form_id)
+            form.errors.append(message)
+        return prop
+
+    def add_property_detail(self, detail_type, root_case_type, module_id, column):
+        field = column.field
+        if field == '#owner_name':
+            return None
+
+        parts = field.split('/')
+        if parts and parts[0] == 'user':
+            root_case_type = USERCASE_TYPE
+            field = field.split('user/')[1]
+
+        try:
+            props = self.get_property_list(root_case_type, field)
+        except CaseMetaException as e:
+            props = [self.add_property_error(root_case_type, field, form_id=None, message=None)]
+            error = six.text_type(e)
+        else:
+            error = None
+        for prop in props:
+            prop.add_detail(detail_type, module_id, column.header, column.useXpathExpression, error)
+
+    def get_error_property(self, case_type, name):
+        type_ = self.get_type(case_type)
+        type_.has_errors = True
+        return type_.get_property(name, allow_parent=True)
+
+    def get_type(self, name):
+        if not name:
+            return CaseTypeMeta(name='')
+
+        try:
+            type_ = next(type_ for type_ in self.case_types if type_.name == name)
+        except StopIteration:
+            type_ = CaseTypeMeta(name=name)
+            self.case_types.append(type_)
+
+        return type_
